+++ test_report_20250908_115557.txt
@@ -1,0 +1,88 @@
+Komoplane Test Report
+Generated on: Mon Sep  8 11:55:57 BST 2025
+==========================
+
+Backend Test Coverage:
+github.com/komodorio/komoplane/main.go:32:					main				0.0%
+github.com/komodorio/komoplane/main.go:85:					parseFlags			0.0%
+github.com/komodorio/komoplane/main.go:111:					setupLogging			0.0%
+github.com/komodorio/komoplane/pkg/backend/api.go:14:				NewRouter			0.0%
+github.com/komodorio/komoplane/pkg/backend/api.go:52:				slowness			0.0%
+github.com/komodorio/komoplane/pkg/backend/api.go:68:				devNoCORS			0.0%
+github.com/komodorio/komoplane/pkg/backend/api.go:75:				errSet500			0.0%
+github.com/komodorio/komoplane/pkg/backend/api.go:86:				configureRoutes			0.0%
+github.com/komodorio/komoplane/pkg/backend/api.go:129:				configureStatic			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:72:			NewManagedUnstructured		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:86:			GetProviderConfigReference	0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:101:			SetProviderConfigReference	0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:105:			GetStatus			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:117:			GetProviders			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:126:			GetProvider			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:135:			GetProviderEvents		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:155:			GetProviderConfigs		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:164:			GetProviderConfigsInner		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:196:			LoadCRDs			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:271:			IsManagedResourceCRD		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:309:			ExtractProviderNameFromCRD	0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:344:			MatchesProviderPattern		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:374:			GetClaims			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:393:			GetClaim			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:442:			fillCompositionByRef		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:481:			getDynamicResource		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:515:			GetManageds			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:553:			getCachedMRDs			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:573:			allMRDs				0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:589:			GetManaged			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:690:			GetManagedNamespaced		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:770:			GetComposites			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:786:			GetCompositions			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:795:			GetComposition			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:812:			GetXRDs				0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:821:			GetEvents			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:840:			GetComposite			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:955:			fillManagedResources		0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:990:			fillXRDList			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:1035:			matchXR				0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:1051:			cachedListXRDs			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:1067:			NewController			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:1120:			durationFromEnv			0.0%
+github.com/komodorio/komoplane/pkg/backend/controller.go:1130:			getK8sConfig			0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/apiV1.go:19:		NewAPIv1Client			0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/apiV1.go:34:		Providers			0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/composition.go:19:	List				0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/composition.go:30:	Get				0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/crds.go:26:		Get				0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/crds.go:54:		List				90.9%
+github.com/komodorio/komoplane/pkg/backend/crossplane/crds.go:77:		getPluralKind			0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/crds.go:97:		NewCRDsClient			0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/events.go:21:		List				0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/events.go:39:		NewEventsClient			0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/extV1.go:21:		NewEXTv1Client			0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/extV1.go:36:		XRDs				0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/extV1.go:57:		Compositions			0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/providers.go:19:		List				0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/providers.go:30:		Get				0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/xrds.go:23:		List				0.0%
+github.com/komodorio/komoplane/pkg/backend/crossplane/xrds.go:71:		Get				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:13:			NewPrefixedFS			0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:26:			Open				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:36:			Close				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:40:			Read				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:44:			Seek				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:48:			Readdir				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:52:			Stat				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:59:			Name				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:63:			Size				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:67:			Mode				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:71:			ModTime				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:75:			IsDir				0.0%
+github.com/komodorio/komoplane/pkg/backend/fs_helper.go:79:			Sys				0.0%
+github.com/komodorio/komoplane/pkg/backend/server.go:25:			StartServer			0.0%
+github.com/komodorio/komoplane/pkg/backend/server.go:56:			startBackgroundServer		0.0%
+github.com/komodorio/komoplane/pkg/backend/server.go:84:			checkUpgrade			0.0%
+github.com/komodorio/komoplane/pkg/backend/utils/utils.go:5:			Plural				100.0%
+total:										(statements)			3.6%
+
+Frontend test coverage report available at: tests/coverage/lcov-report/index.html
+
+Test execution completed successfully
